+++ packages/edit-site/src/components/left-sidebar/navigation-panel/menus/templates-pages.js
@@ -1,0 +1,43 @@
+/**
+ * WordPress dependencies
+ */
+import {
+	__experimentalNavigationGroup as NavigationGroup,
+	__experimentalNavigationMenu as NavigationMenu,
+} from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import TemplateNavigationItems from '../template-navigation-items';
+
+export default function TemplatesPagesMenu( { templates, onActivateItem } ) {
+	const defaultTemplate = templates?.find( ( { slug } ) => slug === 'page' );
+	const specificPageTemplates = templates?.filter( ( { slug } ) =>
+		slug.startsWith( 'page-' )
+	);
+
+	return (
+		<NavigationMenu
+			menu="templates-pages"
+			title="Pages"
+			parentMenu="templates"
+		>
+			<NavigationGroup title="Specific">
+				<TemplateNavigationItems
+					templates={ specificPageTemplates }
+					onActivateItem={ onActivateItem }
+				/>
+			</NavigationGroup>
+
+			{ defaultTemplate && (
+				<NavigationGroup title="General">
+					<TemplateNavigationItems
+						templates={ defaultTemplate }
+						onActivateItem={ onActivateItem }
+					/>
+				</NavigationGroup>
+			) }
+		</NavigationMenu>
+	);
+}