+++ packages/edit-site/src/components/sidebar/global-styles-sidebar.js
@@ -6,14 +6,17 @@ import { omit } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { PanelBody, TabPanel } from '@wordpress/components';
+import { Button, PanelBody, TabPanel } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { getBlockType } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
  */
-import { useGlobalStylesContext } from '../editor/global-styles-provider';
+import {
+	useGlobalStylesContext,
+	useGlobalStylesReset,
+} from '../editor/global-styles-provider';
 import DefaultSidebar from './default-sidebar';
 import { GLOBAL_CONTEXT } from '../editor/utils';
 import TypographyPanel from './typography-panel';
@@ -25,6 +28,7 @@ export default ( { identifier, title, icon, closeLabel } ) => {
 		getStyleProperty,
 		setStyleProperty,
 	} = useGlobalStylesContext();
+	const [ canRestart, onReset ] = useGlobalStylesReset();
 
 	if ( typeof contexts !== 'object' || ! contexts?.[ GLOBAL_CONTEXT ] ) {
 		// No sidebar is shown.
@@ -33,10 +37,25 @@ export default ( { identifier, title, icon, closeLabel } ) => {
 
 	return (
 		<DefaultSidebar
+			className="edit-site-global-styles-sidebar"
 			identifier={ identifier }
 			title={ title }
 			icon={ icon }
 			closeLabel={ closeLabel }
+			header={
+				<>
+					<strong>{ title }</strong>
+					<Button
+						className="edit-site-global-styles-sidebar__reset-button"
+						isSmall
+						isTertiary
+						disabled={ ! canRestart }
+						onClick={ onReset }
+					>
+						{ __( 'Reset to defaults' ) }
+					</Button>
+				</>
+			}
 		>
 			<TabPanel
 				tabs={ [