+++ packages/edit-widgets/src/index.js
@@ -18,8 +18,8 @@ import {
 import './store';
 import './hooks';
 import { create as createLegacyWidget } from './blocks/legacy-widget';
+import * as widgetArea from './blocks/widget-area';
 import EditWidgetsInitializer from './components/edit-widgets-initializer';
-import CustomizerEditWidgetsInitializer from './components/customizer-edit-widgets-initializer';
 
 /**
  * Initializes the block editor in the widgets screen.
@@ -32,27 +32,10 @@ export function initialize( id, settings ) {
 	if ( process.env.GUTENBERG_PHASE === 2 ) {
 		__experimentalRegisterExperimentalCoreBlocks( settings );
 		registerBlock( createLegacyWidget( settings ) );
+		registerBlock( widgetArea );
 	}
 	render(
 		<EditWidgetsInitializer settings={ settings } />,
-		document.getElementById( id )
-	);
-}
-
-/**
- * Initializes the block editor in the widgets Customizer section.
- *
- * @param {string} id       ID of the root element to render the section in.
- * @param {Object} settings Block editor settings.
- */
-export function customizerInitialize( id, settings ) {
-	registerCoreBlocks();
-	if ( process.env.GUTENBERG_PHASE === 2 ) {
-		__experimentalRegisterExperimentalCoreBlocks( settings );
-		registerBlock( createLegacyWidget( settings ) );
-	}
-	render(
-		<CustomizerEditWidgetsInitializer settings={ settings } />,
 		document.getElementById( id )
 	);
 }